worker/pyproject.toml
[project]
name = "vortex-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-redis>=0.61b0",
    "opentelemetry-instrumentation-requests>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.7.0",
    "qdrant-client>=1.16.2",
    "redis==5.0.1",
]
